se crearon los constructores para todas las clases

+++ src/org/cmendoza/pooherenciainterfaz/interfaz/IProducto.java
@@ -1,6 +1,8 @@
 package org.cmendoza.pooherenciainterfaz.interfaz;
 
+import org.cmendoza.pooherenciainterfaz.modelo.Producto;
+
 public interface IProducto {
     int getPrecio();
-    int getPrecioVenta();
+    double getPrecioVenta();
 }
